pyproject.toml
[tool.poetry]
name = "library-snippet-registration"
packages = [
    {include = "snippet"}
]
version = "2.0.0"
description = "Register the code of your own library in the vscode or cursor snippet."
authors = ["sakagami0615 <sakagami0615@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0.1"
tox = "^4.18.0"

[tool.poetry.group.dev.dependencies]
types-PyYAML = "^6.0.12.12"
mypy = "^1.9.0"
ruff = "^0.3.2"
freezegun = "^1.4.0"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
check_untyped_defs = true
explicit_package_bases = true
mypy_path = "$MYPY_CONFIG_FILE_DIR"
namespace_packages = true

[tool.ruff]
target-version = "py311"
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".git",
]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["C", "E", "F", "W", "I", "D"]
ignore = ["D100", "D104", "D105", "D107", "D415"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-single-line = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
